% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{ancestor}
\alias{ancestor}
\title{ancestor}
\usage{
ancestor(.data, ...)
}
\arguments{
\item{.data}{A tree_tbl data frame}

\item{...}{additional parameter}
}
\value{
tree_tbl data frame
}
\description{
access ancestor data
}
\author{
guangchuang yu
}
