% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/converter.R,
%   R/method-as-treedata.R
\name{as.treedata}
\alias{as.treedata}
\alias{as.treedata.tbl_tree}
\alias{as.treedata.phylo}
\title{as.treedata}
\usage{
as.treedata(tree, ...)

\method{as.treedata}{tbl_tree}(tree, ...)

\method{as.treedata}{phylo}(tree, boot = NULL, ...)
}
\arguments{
\item{tree}{input tree, a \code{phylo} object}

\item{...}{additional parameters}

\item{boot}{optional, can be bootstrap value from ape::boot.phylo}
}
\value{
treedata object
}
\description{
convert a tree object to treedata object

convert phylo to treedata
}
\details{
converting phylo object to treedata object
}
\examples{
library(ape)
set.seed(2017)
tree <- rtree(4)
d <- tibble(label = paste0('t', 1:4),
           trait = rnorm(4))
x <- as_tibble(tree)
full_join(x, d, by = 'label') \%>\% as.treedata
}
\author{
Guangchuang Yu
}
