% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/get-data.R
\docType{methods}
\name{get.data}
\alias{get.data}
\alias{get.data,treedata-method}
\title{get.data method}
\usage{
get.data(object, ...)

\S4method{get.data}{treedata}(object)
}
\arguments{
\item{object}{\code{treedata} object}

\item{...}{additional parameter}
}
\value{
associated data of phylogeny
}
\description{
get.data method
}
