% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-groupOTU.R
\name{getNodeNum}
\alias{getNodeNum}
\title{getNodeNum}
\usage{
getNodeNum(tree)
}
\arguments{
\item{tree}{tree object}
}
\value{
number
}
\description{
calculate total number of nodes
}
\examples{
getNodeNum(rtree(30))
}
\author{
Guangchuang Yu
}
