% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{groupClade}
\alias{groupClade}
\title{groupClade}
\usage{
groupClade(.data, .node, group_name = "group", overlap = "overwrite", ...)
}
\arguments{
\item{.data}{tree object (phylo, treedata, tbl_tree, ggtree etc.)}

\item{.node}{selected nodes}

\item{group_name}{character the name of the group cluster, default is \code{group}.}

\item{overlap}{character one of \code{overwrite},\code{origin} and \code{abandon},
default is \code{overwrite}.}

\item{...}{additional parameter}
}
\value{
updated tree with group information or group index
}
\description{
grouping clades
}
\author{
Guangchuang Yu
}
