% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{groupOTU}
\alias{groupOTU}
\title{groupOTU}
\usage{
groupOTU(.data, .node, group_name = "group", ...)
}
\arguments{
\item{.data}{tree object (phylo, treedata, tbl_tree, ggtree etc.)}

\item{.node}{selected nodes}

\item{group_name}{character the name of the group cluster, default is \code{group}.}

\item{...}{additional parameter}
}
\value{
updated tree with group information or group index
}
\description{
grouping OTUs
}
\author{
Guangchuang Yu
}
