% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{mrca}
\alias{mrca}
\title{mrca}
\usage{
mrca(.data, ...)
}
\arguments{
\item{.data}{A tbl_tree data frame}

\item{...}{additional parameters}
}
\value{
tbl_tree data frame
}
\description{
access most recent common ancestor data
}
\author{
guangchuang yu
}
