% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-accessor.R
\name{node.label}
\alias{node.label}
\title{extract the node label of phylo, treedata or tbl_tree}
\usage{
node.label(x, node = "internal", ...)
}
\arguments{
\item{x}{object, should be one of \code{treedata}, \code{phylo} or \code{tbl_tree}.}

\item{node}{character, to extract which type node label,
default is \code{internal}, should be one of \code{internal},
\code{external}, \code{all}, \code{tip}.}

\item{...}{additional parameters.}
}
\value{
label character vector.
}
\description{
extract the node label of phylo, treedata or tbl_tree
}
