% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{nodeid}
\alias{nodeid}
\title{nodeid}
\usage{
nodeid(tree, label)
}
\arguments{
\item{tree}{tree object}

\item{label}{tip/node label(s)}
}
\value{
node number
}
\description{
convert tree label to internal node number
}
\author{
Guangchuang Yu
}
