% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{nodelab}
\alias{nodelab}
\title{nodelab}
\usage{
nodelab(tree, id)
}
\arguments{
\item{tree}{tree object}

\item{id}{node number}
}
\value{
tip/node label(s)
}
\description{
convert internal node number tip/node label
}
\author{
Guangchuang Yu
}
