% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/ancestor.R
\name{parent}
\alias{parent}
\alias{parent.tbl_tree}
\title{parent}
\usage{
parent(.data, .node, ...)

\method{parent}{tbl_tree}(.data, .node, ...)
}
\arguments{
\item{.data}{phylo or tbl_tree object}

\item{.node}{node number}

\item{...}{additional parameters}
}
\value{
parent data
}
\description{
access parent data
}
\examples{
library(ape)
tree <- rtree(4)
x <- as_tibble(tree)
parent(x, 2)
}
\author{
Guangchuang Yu
}
