% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/ape.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as.phylo}
\alias{Nnode}
\alias{rtree}
\alias{read.tree}
\alias{Ntip}
\alias{is.rooted}
\alias{root}
\alias{\%>\%}
\alias{\%<>\%}
\alias{as_tibble}
\alias{tibble}
\alias{filter}
\alias{arrange}
\alias{select}
\alias{rename}
\alias{mutate}
\alias{transmute}
\alias{summarise}
\alias{summarize}
\alias{full_join}
\alias{.data}
\alias{left_join}
\alias{pull}
\alias{unnest}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ape}{\code{\link[ape:as.phylo]{as.phylo()}}, \code{\link[ape:as.phylo]{as.phylo()}}, \code{\link[ape:is.rooted]{is.rooted()}}, \code{\link[ape:Nnode]{Nnode()}}, \code{\link[ape:Ntip]{Ntip()}}, \code{\link[ape:read.tree]{read.tree()}}, \code{\link[ape:root]{root()}}, \code{\link[ape:rtree]{rtree()}}}

  \item{dplyr}{\code{\link[dplyr:arrange]{arrange()}}, \code{\link[dplyr:filter]{filter()}}, \code{\link[dplyr:full_join]{full_join()}}, \code{\link[dplyr:left_join]{left_join()}}, \code{\link[dplyr:mutate]{mutate()}}, \code{\link[dplyr:pull]{pull()}}, \code{\link[dplyr:rename]{rename()}}, \code{\link[dplyr:rename]{rename()}}, \code{\link[dplyr:select]{select()}}, \code{\link[dplyr:summarise]{summarise()}}, \code{\link[dplyr:summarize]{summarize()}}, \code{\link[dplyr:transmute]{transmute()}}}

  \item{magrittr}{\code{\link[magrittr:\%<>\%]{\%<>\%}}, \code{\link[magrittr:\%>\%]{\%>\%}}}

  \item{rlang}{\code{\link[rlang:.data]{.data}}}

  \item{tibble}{\code{\link[tibble:as_tibble]{as_tibble()}}, \code{\link[tibble:tibble]{tibble()}}}

  \item{tidyr}{\code{\link[tidyr:unnest]{unnest()}}}
}}

