% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{root}
\alias{root}
\title{root}
\usage{
root(.data, ...)
}
\arguments{
\item{.data}{A tbl_tree data frame}

\item{...}{additional parameters}
}
\value{
tbl_tree data frame
}
\description{
access root data
}
\author{
guangchuang yu
}
