% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-package.R
\docType{package}
\name{tidytree-package}
\alias{tidytree}
\alias{tidytree-package}
\title{tidytree: A Tidy Tool for Phylogenetic Tree Data Manipulation}
\description{
Phylogenetic trees contain components such as nodes, edges, branches, and associated data. As the core data manipulation tool in the 'ggtree' ecosystem, 'tidytree' provides tools for converting tree objects to tidy data frames and tidy interfaces for manipulating, analyzing, and visualizing tree data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://yulab-smu.top/contribution-tree-data/}
  \item Report bugs at \url{https://github.com/YuLab-SMU/tidytree/issues}
}

}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com} (\href{https://orcid.org/0000-0002-6485-8781}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Guangchuang Yu \email{guangchuangyu@gmail.com} (\href{https://orcid.org/0000-0002-6485-8781}{ORCID}) [copyright holder]
}

Other contributors:
\itemize{
  \item Bradley Jones \email{brj1@sfu.ca} [contributor]
  \item Zebulun Arendsee \email{zbwrnz@gmail.com} [contributor]
}

}
\keyword{internal}
