% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-accessor.R
\name{tip.label}
\alias{tip.label}
\title{extract the tip label of phylo treedata or tbl_tree}
\usage{
tip.label(x, ...)
}
\arguments{
\item{x}{object, should be one of \code{treedata}, \code{phylo} or \code{tbl_tree}.}

\item{...}{additional parameters.}
}
\description{
extract the tip label of phylo treedata or tbl_tree
}
