% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/ancestor.R
\name{ancestor}
\alias{ancestor}
\alias{ancestor.tbl_tree}
\title{ancestor}
\usage{
ancestor(.data, .node, ...)

\method{ancestor}{tbl_tree}(.data, .node, ...)
}
\arguments{
\item{.data}{phylo or tbl_tree object}

\item{.node}{node number}

\item{...}{additional parameters}
}
\value{
ancestor data
}
\description{
access ancestor data
}
\examples{
library(ape)
tree <- rtree(4)
x <- as_data_frame(tree)
ancestor(x, 3)
}
\author{
guangchuang yu
}
