% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{as.treedata}
\alias{as.treedata}
\title{as.treedata}
\usage{
as.treedata(tree, ...)
}
\arguments{
\item{tree}{tree object}

\item{...}{additional parameters}
}
\value{
treedata object
}
\description{
convert a tree object to treedata object
}
