% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/offspring.R
\name{offspring}
\alias{offspring}
\alias{offspring.tbl_tree}
\title{offspring}
\usage{
offspring(.data, .node, ...)

\method{offspring}{tbl_tree}(.data, .node, ...)
}
\arguments{
\item{.data}{phylo or tbl_tree object}

\item{.node}{node number}

\item{...}{additional parameters}
}
\value{
offspring data
}
\description{
access offspring data
}
\examples{
library(ape)
tree <- rtree(4)
x <- as_data_frame(tree)
offspring(x, 4)
}
\author{
guangchuang yu
}
