% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{offspring}
\alias{offspring}
\title{offspring}
\usage{
offspring(.data, ...)
}
\arguments{
\item{.data}{A tree_tbl data frame}

\item{...}{additional parameter}
}
\value{
tree_tbl data frame
}
\description{
access offspring data
}
\author{
guangchuang yu
}
