% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{parent}
\alias{parent}
\title{parent}
\usage{
parent(.data, ...)
}
\arguments{
\item{.data}{A tree_tbl data frame}

\item{...}{additional parameter}
}
\value{
tree_tbl data frame
}
\description{
access parent data
}
\author{
guangchuang yu
}
