% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as.phylo}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{\%<>\%}
\alias{reexports}
\alias{as_data_frame}
\alias{reexports}
\alias{data_frame}
\alias{reexports}
\alias{tibble}
\alias{reexports}
\alias{filter}
\alias{reexports}
\alias{arrange}
\alias{reexports}
\alias{select}
\alias{reexports}
\alias{rename}
\alias{reexports}
\alias{mutate}
\alias{reexports}
\alias{transmute}
\alias{reexports}
\alias{summarise}
\alias{reexports}
\alias{summarize}
\alias{reexports}
\alias{full_join}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ape}{\code{\link[ape]{as.phylo}}}

  \item{dplyr}{\code{\link[dplyr]{filter}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{transmute}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{summarize}}, \code{\link[dplyr]{full_join}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}, \code{\link[magrittr]{\%<>\%}}}

  \item{tibble}{\code{\link[tibble]{as_data_frame}}, \code{\link[tibble]{data_frame}}, \code{\link[tibble]{tibble}}}
}}

