% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{reexports}
\alias{\%<>\%}
\alias{reexports}
\alias{as_data_frame}
\alias{reexports}
\alias{filter}
\alias{reexports}
\alias{arrange}
\alias{reexports}
\alias{select}
\alias{reexports}
\alias{rename}
\alias{reexports}
\alias{mutate}
\alias{reexports}
\alias{transmute}
\alias{reexports}
\alias{summarise}
\alias{reexports}
\alias{groupOTU}
\alias{reexports}
\alias{groupClade}
\alias{reexports}
\alias{rootnode}
\alias{reexports}
\alias{rootnode}
\alias{reexports}
\alias{sibling}
\alias{reexports}
\alias{offspring}
\alias{reexports}
\alias{parent}
\alias{reexports}
\alias{ancestor}
\alias{reexports}
\alias{as.treedata}
\alias{reexports}
\alias{as.phylo}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{filter}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{transmute}}, \code{\link[dplyr]{summarise}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}, \code{\link[magrittr]{\%<>\%}}}

  \item{tibble}{\code{\link[tibble]{as_data_frame}}}

  \item{treeio}{\code{\link[treeio]{groupOTU}}, \code{\link[treeio]{groupClade}}, \code{\link[treeio]{rootnode}}, \code{\link[treeio]{rootnode}}, \code{\link[treeio]{sibling}}, \code{\link[treeio]{offspring}}, \code{\link[treeio]{parent}}, \code{\link[treeio]{ancestor}}, \code{\link[treeio]{as.treedata}}, \code{\link[treeio]{as.phylo}}}
}}

