% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as.phylo}
\alias{\%>\%}
\alias{\%<>\%}
\alias{as_tibble}
\alias{tibble}
\alias{filter}
\alias{arrange}
\alias{select}
\alias{rename}
\alias{mutate}
\alias{transmute}
\alias{summarise}
\alias{summarize}
\alias{full_join}
\alias{.data}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ape}{\code{\link[ape]{as.phylo}}}

  \item{dplyr}{\code{\link[dplyr]{arrange}}, \code{\link[dplyr]{filter}}, \code{\link[dplyr:mutate-joins]{full_join}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{rename}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr:summarise]{summarize}}, \code{\link[dplyr:mutate]{transmute}}}

  \item{magrittr}{\code{\link[magrittr:compound]{\%<>\%}}, \code{\link[magrittr:pipe]{\%>\%}}}

  \item{rlang}{\code{\link[rlang:tidyeval-data]{.data}}}

  \item{tibble}{\code{\link[tibble]{as_tibble}}, \code{\link[tibble]{tibble}}}
}}

