% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytree-package.R
\docType{package}
\name{tidytree-package}
\alias{tidytree}
\alias{tidytree-package}
\title{tidytree: What the Package Does (one line, title case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
